[BACKLOG-5949] - Setting the bundle context as part of all karaf types (server, client, ...) in the activator.
This is required to allow kettle plugins to use PentahoSystem to register objects into the OSGi container and register them as services. In this particular case, it allows the step analyzers for kettle xml steps in the new plugin to contribute to a service-list used by the metaverse for lineage.